"Enable prefab debugger" option
