+++ QModManager/Prefab Debugger/PrefabDebugger.cs
@@ -132,7 +132,7 @@ namespace QModManager.Debugger
 
         private void LateUpdate()
         {
-            if ((Input.GetKeyDown(KeyCode.F9)) || (showDebugger && Input.GetKeyDown(KeyCode.Escape)))
+            if ((Input.GetKeyDown(KeyCode.F9) && PlayerPrefs.GetInt("QModManager_PrefabDebugger_Enable") == 1) || (showDebugger && Input.GetKeyDown(KeyCode.Escape)))
             {
                 showDebugger = !showDebugger;
                 UWE.Utils.alwaysLockCursor = !showDebugger;
